File bucket

+++ frontend/src/views/Forum/Forum.jsx
@@ -28,7 +28,6 @@ const TOPICS_DISC = 'topicos_disc/';
 const TOPIC = 'topicos/';
 const FILES = 'arquivos/';
 const FILESTOPIC = 'arquivos_topic/';
-const FILEFOLDER = 'http://localhost:80/uploads/';
 const MAXFILES = 5;
 const MAXFILESIZE = 5000000; //5MB
 
@@ -50,11 +49,9 @@ class Forum extends React.Component {
       files: [], 
       prevFiles: [],
       arquivo: {
-        nome: '',
+        name: '',
         upload: '',
         topico_pai: '',
-        formato: '',
-        tamanho: '',
       }
     };
     this.handleInputChange = this.handleInputChange.bind(this);
@@ -73,7 +70,7 @@ class Forum extends React.Component {
       const topicId = this.props.search.split('=')[1];   
       const urlTopic = API + TOPIC + topicId + '/';
       const urlFiles = API + FILESTOPIC + this.props.search;
-      // console.log(url); 
+       
       fetch(urlTopic, {
         headers: {
           'Authorization': 'Bearer ' + this.Auth.getToken()
@@ -98,28 +95,9 @@ class Forum extends React.Component {
       })
         .then(response => response.json())
         .then(data => {
-          data.forEach(file => {
-            let splited = file.upload.split('/');
-            let path = splited[splited.length - 1];
-            const fileId = file.id;
-            console.log(FILEFOLDER + path);    
-            // Access-Control-Allow-Origin: '*' Header is needed
-            fetch(FILEFOLDER + path, {
-              headers: {
-                'Authorization': 'Bearer ' + this.Auth.getToken()
-              }
-            })
-              .then(response => response.blob())  
-              .then(blob => {     
-                let fileObject = new File([blob], path);              
-                let url = window.URL.createObjectURL(blob);         
-                fileObject['id'] = fileId;              
-                fileObject['url'] = url;              
-                this.setState(prevState => ({            
-                  files: [...prevState.files, fileObject],
-                  prevFiles: [...prevState.prevFiles, fileObject],
-                }))
-              })
+          this.setState ({
+            files: data,
+            prevFiles: data,
           });
         });
     }
@@ -133,7 +111,6 @@ class Forum extends React.Component {
     let selTopic = {...this.state.selTopic};
     selTopic[name] = value; 
     this.setState({selTopic});  
-    // console.log(this.state.selTopic);  
 
   }
 
@@ -190,13 +167,10 @@ class Forum extends React.Component {
     if ((newFilesList.length) > MAXFILES) {
       this.setState({isFullOfFiles: true});       
     } else {
-      newFilesList.forEach(file => {     
-        let url = window.URL.createObjectURL(file);
-        file['url'] = url; 
+      newFilesList.forEach(file => {             
         this.setState({ files: newFilesList });            
       });        
     }
-    console.log(this.state.files);
   }
 
   updateTopic = () => {
@@ -217,16 +191,15 @@ class Forum extends React.Component {
       })    
       .then(response => response.json())
       .then(data => {
-          console.log(data);          
-          this.state.files.forEach(function(f) {
-            if (!f['id']) {
+          var fileArray;
+          this.state.files.forEach((f, index) => {
+            if (!f['id'] && this.Auth.loggedIn()) {
               let data = new FormData();
               let file = new File([f.slice(0,-1)], Date.now() + '___' + f.name, {type: f.type}); // Grants that the names will be different   
-              data.append("nome", file.name);  
+              data.append("name", file.name);  
               data.append("upload", file);             
-              data.append("topico_pai", topico.id);             
-              data.append("formato", file.type);             
-              data.append("tamanho", file.size);             
+              data.append("topico_pai", topico.id);                        
+              data.append("size", file.size);                        
                       
               fetch(API + FILES, {
                 method: 'post',
@@ -237,10 +210,13 @@ class Forum extends React.Component {
                 },
               })    
               .then(response => response.json())
-              .then(data => {            
-                  console.log(data);                 
+              .then(data => { 
+                let filesOld = this.state.files;
+                filesOld.splice(index, 1);
+                filesOld.push(data);
+                this.setState({ filesOld })                                             
               });
-            } 
+            }             
           }); 
           console.log(this.state.prevFiles);
           this.state.prevFiles.forEach((prevF, index) => {
@@ -250,7 +226,7 @@ class Forum extends React.Component {
                 deletedFile = false;                
               }
             })
-            if (deletedFile) {
+            if (deletedFile && this.Auth.loggedIn()) {
               fetch(API + FILES + prevF.id + '/', {
                 method: 'delete',
                 headers: {
@@ -258,8 +234,6 @@ class Forum extends React.Component {
                 }
               })    
               .then(response => {
-                console.log(response);           
-                console.log("data"); 
                 console.log("deleted:");
                 console.log(prevF);                 
               });              
@@ -288,6 +262,10 @@ class Forum extends React.Component {
     this.setState({ dialogOpen: false });
   };
 
+  downloadFile(url) {
+    window.open(url);
+  }
+
   truncate(string, maxSize) {        
     if (string.length > maxSize) {
       return (string.substring(0,maxSize) + "...");
@@ -297,9 +275,6 @@ class Forum extends React.Component {
   }
 
   render() {   
-    // console.log(this.state.selTopic);
-    // console.log(this.props.search);
-    // console.log("Files: ");
     let disciplina = this.state.discPai;
     let selTopic = this.state.selTopic;      
     let commentaryList;
@@ -357,7 +332,7 @@ class Forum extends React.Component {
                               rejectClassName="dropReject" 
                               onDrop={this.onDrop.bind(this)}>
                               <InputLabel>Anexos</InputLabel>
-                              <p>Arraste e solte aqui os arquivos que deseja anexar.</p>
+                              <p>Clique aqui ou arraste e solte os arquivos que deseja anexar.</p>
                             </Dropzone>
                           </div>
                           <aside className="filesDiv">
@@ -381,10 +356,10 @@ class Forum extends React.Component {
                                   <li key={f.name}>
                                     <div>
                                       <Icon className='fileIcon'>attachment</Icon>                                  
-                                      <a className='fileName' href={f.url} download={f.name.split('___')[1]}>
+                                      <a className='fileName' onClick={() => (this.downloadFile(f.upload))} download={f.name.split('___')[1]}>
                                         {(f.name.split('___')[1] ? this.truncate( f.name.split('___')[1], 20) : this.truncate( f.name, 20)) + ' - ' + Math.round(f.size/1000) + ' Kbytes'}
                                       </a>
-                                      <a className='downloadFileIcon' href={f.url} download={f.name.split('___')[1]}>
+                                      <a className='downloadFileIcon' onClick={() => (this.downloadFile(f.upload))} download={f.name.split('___')[1]}>
                                         <Icon>cloud_download</Icon>
                                       </a>
                                       <Icon onClick={() => (this.deleteFile(f))} className='delFileIcon'>close</Icon>
@@ -427,10 +402,10 @@ class Forum extends React.Component {
                                       <li key={f.name}>
                                         <div>
                                           <Icon className='fileIcon'>attachment</Icon>                                  
-                                          <a className='fileName' href={f.url} download={f.name.split('___')[1]}>
+                                          <a className='fileName' onClick={() => (this.downloadFile(f.upload))} download={f.name.split('___')[1]}>
                                             {(f.name.split('___')[1] ? this.truncate( f.name.split('___')[1], 20) : this.truncate( f.name, 20)) + ' - ' + Math.round(f.size/1000) + ' Kbytes'}
                                           </a>
-                                          <a className='downloadFileIcon' href={f.url} download={f.name.split('___')[1]}>
+                                          <a className='downloadFileIcon' onClick={() => (this.downloadFile(f.upload))} download={f.name.split('___')[1]}>
                                             <Icon>cloud_download</Icon>
                                           </a>                      
                                         </div>
